requirements.txt
discord.py
dislash.py